requirements.txt
langchain==0.2.17
langchain-openai==0.1.25
langchain-community==0.2.19
tavily-python==0.7.11
python-dotenv==1.0.1